src/main/java/com/example/jdstreetwear/controller/CategoryController.java
package com.example.jdstreetwear.controller;

import com.example.jdstreetwear.model.Category;
import com.example.jdstreetwear.service.CategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;
import java.util.Optional;

@RestController
@RequestMapping("/api/categories")
public class CategoryController {

    @Autowired
    private CategoryService categoryService;

    @GetMapping
    public List<Category> getAllCategories() {
        System.out.println("Received request to get all categories");
        return categoryService.getAllCategories();
    }

    @GetMapping("/{id}")
    public ResponseEntity<Category> getCategoryById(@PathVariable(value = "id") Long id) {
        System.out.println("Received request to get category by ID: " + id);
        return categoryService.getCategoryById(id)
                .map(category -> ResponseEntity.ok().body(category))
                .orElse(ResponseEntity.notFound().build());
    }

    @PostMapping
    public Category createCategory(@RequestBody Category category) {
        System.out.println("Received request to create category: " + category);
        return categoryService.createCategory(category);
    }

    @PutMapping("/{id}")
    public ResponseEntity<Category> updateCategory(@PathVariable(value = "id") Long id, @RequestBody Category categoryDetails) {
        System.out.println("Received request to update category: " + id + " with details: " + categoryDetails);
        return ResponseEntity.ok(categoryService.updateCategory(id, categoryDetails));
    }

    @DeleteMapping("/{id}")
    public ResponseEntity<Void> deleteCategory(@PathVariable(value = "id") Long id) {
        System.out.println("Received request to delete category: " + id);
        categoryService.deleteCategory(id);
        return ResponseEntity.ok().build();
    }
}
